Conflicts:
	src/main/java/com/example/ContainerTerminal/controllers/UserController.java
	src/main/java/com/example/ContainerTerminal/models/Container.java
	src/main/resources/static/script/datepicker.js
	src/main/webapp/WEB-INF/jsp/Payment.jsp
	src/main/webapp/WEB-INF/jsp/welcome.jsp

+++ src/main/java/com/example/ContainerTerminal/models/Seawaybill.java
@@ -30,7 +30,6 @@ import javax.xml.bind.annotation.XmlTransient;
  * @author mario
  */
 @Entity
-
 @Table(name = "seawaybill")
 @XmlRootElement
 @NamedQueries({
@@ -55,7 +54,7 @@ public class Seawaybill implements Serializable {
     @Basic(optional = false)
     @NotNull
     @Column(name = "custom")
-    private short custom;
+    private Integer custom;
     @Basic(optional = false)
     @NotNull
     @Column(name = "paid")
@@ -73,7 +72,7 @@ public class Seawaybill implements Serializable {
         this.id = id;
     }
 
-    public Seawaybill(Integer id, String bookingnumber, short custom, short paid) {
+    public Seawaybill(Integer id, String bookingnumber, Integer custom, short paid) {
         this.id = id;
         this.bookingnumber = bookingnumber;
         this.custom = custom;
@@ -96,11 +95,11 @@ public class Seawaybill implements Serializable {
         this.bookingnumber = bookingnumber;
     }
 
-    public short getCustom() {
+    public Integer getCustom() {
         return custom;
     }
 
-    public void setCustom(short custom) {
+    public void setCustom(Integer custom) {
         this.custom = custom;
     }
